Remove default value on `parseReactElement` options parameters
