+++ src/index.js
@@ -1,14 +1,40 @@
 /* @flow */
 
 import { Element } from 'react';
+import { defaultFunctionValue } from './formatter/formatPropValue';
 import formatTree from './formatter/formatTree';
 import parseReactElement from './parser/parseReactElement';
 import type { Options } from './options';
 
-const reactElementToJsxString = (element: Element<any>, options: Options) => {
+const reactElementToJsxString = (
+  element: Element<any>,
+  {
+    filterProps = [],
+    showDefaultProps = true,
+    showFunctions = false,
+    functionValue = defaultFunctionValue,
+    tabStop = 2,
+    useBooleanShorthandSyntax = true,
+    sortProps = true,
+    maxInlineAttributesLineLength,
+    displayName,
+  }: Options = {}
+) => {
   if (!element) {
     throw new Error('react-element-to-jsx-string: Expected a ReactElement');
   }
+
+  const options = {
+    filterProps,
+    showDefaultProps,
+    showFunctions,
+    functionValue,
+    tabStop,
+    useBooleanShorthandSyntax,
+    sortProps,
+    maxInlineAttributesLineLength,
+    displayName,
+  };
 
   return formatTree(parseReactElement(element, options), options);
 };
